Remove mac builds, fix windows and linux builds
Can't test the mac builds in any case, but apparently the working
directory isn't very well defined (based on a quick web search), and the
engine really expects resources.db to be an actual file which exists in
the current working directory. In the case of the web, I can test it, so
I will do something more specific.